+++ docs/src/examples/components/RadioGroup/Types/RadioGroupColorPickerExample.shorthand.tsx
@@ -17,32 +17,17 @@ class RadioGroupColorPickerExample extends React.Component {
             value: color,
             name: color,
             'aria-label': color,
-            icon: this.createIcon(color),
+            icon: {
+              name: 'stardust-circle',
+              variables: {
+                color,
+              },
+            },
           }))}
           checkedValueChanged={(e, props) => this.setState({ selectedValue: props.value })}
         />
       </div>
     )
-  }
-
-  createIcon(value) {
-    const { selectedValue } = this.state
-    const isSelected = selectedValue === value
-
-    return {
-      styles: {
-        backgroundColor: value,
-        border: '1px solid white',
-        ...(isSelected && {
-          border: '1px solid black',
-        }),
-        borderRadius: '3px',
-        ...(isSelected && {
-          backgroundClip: 'content-box',
-          padding: '5px',
-        }),
-      },
-    }
   }
 }